feat(many): github aware, splitting out redux, refactoring
