+++ src/redux/templates/components/ConnectRedux/ConnectRedux.js
@@ -1,0 +1,12 @@
+import { connect } from 'react-redux';
+import { createSelector } from 'reselect';
+
+import actions from '../../redux/actions';
+
+export default connect(
+	createSelector(
+		(state) => state.entities,
+		(entities) => entities,
+	),
+	actions
+);
