+++ src/redux/templates/entities/foos/schema.normalizr.js
