+++ src/redux/templates/redux/actions.js
