+++ src/redux/templates/redux/entities.actions.js.ejs
