+++ src/redux/templates/redux/entities.reducer.js
