+++ src/server/index.js
@@ -14,7 +14,7 @@ export default class ServerGenerator extends Base {
 					default: 'heroku',
 					choices: [
 						{ name: 'Heroku', value: 'heroku' },
-						{ name: 'Don\'t worry about it', value: null },
+						{ name: 'None', value: null },
 					],
 				}]);
 			})
@@ -26,22 +26,29 @@ export default class ServerGenerator extends Base {
 	}
 	configuring() {
 		this.fs.copy(this.templatePath('Procfile'), this.destinationPath('Procfile'));
+
 		this.npmInstall(
 			[
 				'babel-cli',
 			],
 			{ save: true }
 		);
-
+	}
+	writing() {
 		this.fs.copy(this.templatePath('report/index.js'), this.destinationPath('report/index.js'));
-		this.npmInstall(
-			[
-				'raven',
-			],
-			{ save: true }
+		this.fs.copyTpl(this.templatePath('index.js.ejs'), this.destinationPath('index.js'), Object.assign({}, this, this.options));
+		this.fs.write(
+			this.destinationPath('Procfile.dev'),
+			[this.fs.read(this.templatePath('Procfile.dev')), this.fs.read(this.destinationPath('Procfile.dev'))].join('\n').replace(/\n{2,}/, '\n')
 		);
 
-		this.fs.copyTpl(this.templatePath('index.js.ejs'), this.destinationPath('index.js'), Object.assign({}, this, this.options));
+		this.npmInstall(
+			[
+				'nodemon',
+			],
+			{ saveDev: true }
+		);
+
 		this.npmInstall(
 			[
 				'body-parser',
@@ -49,9 +56,11 @@ export default class ServerGenerator extends Base {
 				'feathers',
 				'feathers-errors',
 				'helmet',
+				'raven',
 			],
 			{ save: true }
 		);
+
 		if (this.options.render === 'server') {
 			this.npmInstall(
 				[
@@ -60,18 +69,6 @@ export default class ServerGenerator extends Base {
 				{ save: true }
 			);
 		}
-	}
-	writing() {
-		this.fs.write(
-			this.destinationPath('Procfile.dev'),
-			[this.fs.read(this.templatePath('Procfile.dev')), this.fs.read(this.destinationPath('Procfile.dev'))].join('\n').replace(/\n+/, '\n')
-		);
-		this.npmInstall(
-			[
-				'nodemon',
-			],
-			{ saveDev: true }
-		);
 	}
 	_prompt(prompts) {
 		return this.prompt(prompts).then((answers) => {
