+++ src/web/templates/components/App/App.js.ejs
@@ -1,19 +1,13 @@
 import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import { createStructuredSelector } from 'reselect';
 
-import actions from '../../redux/actions';
 import styles from './App.styles';
 
-export default connect(
-	createStructuredSelector({
-		foos: (state) => state.entities.foos,
-	}),
-	actions
-)(class App extends Component {
+export default class App extends Component {
 	static propTypes = {
 		children: React.PropTypes.node.isRequired,
+<%_ if (state === 'redux') { _%>
 		foos:     React.PropTypes.object.isRequired,
+<%_ } _%>
 	}
 	render() {
 		return (
@@ -22,4 +16,4 @@ export default connect(
 			</div>
 		);
 	}
-});
+}
