+++ src/web/templates/components/routes.js.ejs
@@ -1,0 +1,15 @@
+import IndexRoute from 'react-router/lib/IndexRoute';
+import React from 'react';
+import Route from 'react-router/lib/Route';
+
+import App from './App/App';
+<%_ if (state === 'redux') { _%>
+import ConnectRedux from './ConnectRedux/ConnectRedux';
+<%_ } _%>
+import Foo from './Foo/Foo';
+
+export default (
+	<Route path="/" component={<% if (state === 'redux') { %>ConnectRedux(App)<% } else { %>App<% } %>}>
+		<IndexRoute component={<% if (state === 'redux') { %>ConnectRedux(Foo)<% } else { %>Foo<% } %>} />
+	</Route>
+);
