+++ src/web/templates/index.web.js.ejs
@@ -1,12 +1,17 @@
 import './report';
 import React from 'react';
+<%_ if (state === 'redux') { _%>
 import { Provider } from 'react-redux';
+<%_ } _%>
 import { Router, browserHistory } from 'react-router';
 import { render } from 'react-dom';
 
 import './components/global.styles';
+<%_ if (state === 'redux') { _%>
 import createStore from './redux/createStore';
+<%_ } _%>
 import routes from './components/routes';
+<%_ if (state === 'redux') { _%>
 
 const stateElement = document.getElementById('state');
 let state;
@@ -23,12 +28,21 @@ if (stateElement) {
 		}
 	}
 }
+<%_ } _%>
+
 render(
+<%_ if (state === 'redux') { _%>
 	<Provider store={createStore(state)}>
 		<Router history={browserHistory} routes={routes} onUpdate={global.ga && function() {
 			global.ga('set', 'page', global.location.pathname + global.location.search + global.location.hash);
 			global.ga('send', 'pageview');
 		}} />
 	</Provider>,
+<%_ } else { _%>
+	<Router history={browserHistory} routes={routes} onUpdate={global.ga && function() {
+		global.ga('set', 'page', global.location.pathname + global.location.search + global.location.hash);
+		global.ga('send', 'pageview');
+	}} />,
+<%_ } _%>
 	global.document.getElementById('mount')
 );
