+++ src/web/templates/webpack.config.babel.js.ejs
@@ -8,7 +8,7 @@ import CleanWebpackPlugin from 'clean-webpack-plugin';
 import ExtractTextPlugin from 'extract-text-webpack-plugin';
 import FaviconsWebpackPlugin from 'favicons-webpack-plugin';
 <%_ if (render === 'web') { _%>
-import Helmet from "react-helmet";
+import Helmet from 'react-helmet';
 import React from 'react';
 <%_ } _%>
 import HtmlWebpackPlugin from 'html-webpack-plugin';
@@ -26,15 +26,19 @@ import reactRouterToArray from 'react-router-to-array';
 import tail from 'lodash.tail';
 import webpack from 'webpack';
 <%_ if (render === 'web') { _%>
-import { GeneralError, NotFoundError } from 'feathers-errors';
+import { GeneralError } from 'feathers-errors';
+<%_ if (state === 'redux') { _%>
 import { Provider } from 'react-redux';
+<%_ } _%>
 import { RouterContext, match } from 'react-router';
 import { renderToString } from 'react-dom/server';
 <%_ } _%>
 <%_ if (render !== 'server') { _%>
 
 	<%_ if (render === 'web') { _%>
+		<%_ if (state === 'redux') { _%>
 import createStore from './redux/createStore';
+		<%_ } _%>
 	<%_ } _%>
 import routes from './components/routes';
 <%_ } _%>
@@ -116,15 +120,13 @@ export default {
 <%_ if (render !== 'server') { _%>
 		.concat(reactRouterToArray(routes).map((route) => {
 	<%_ if (render === 'web') { _%>
-			let error            = 'nope';
-			let redirectLocation = 'nope';
-			let renderProps      = 'nope';
+			let error       = 'nope';
+			let renderProps = 'nope';
 			match({ routes: routes, location: route }, (matchError, matchRedirectLocation, matchRenderProps) => {
 				// HACK Using react-router's zalgo-ed match
 				// - Discussion about it being zalgo-ed: https://github.com/reactjs/react-router/issues/1777#issuecomment-148415982
 				// - Why it's bad: http://blog.izs.me/post/59142742143/designing-apis-for-asynchrony
 				error = matchError;
-				redirectLocation = matchRedirectLocation;
 				renderProps = matchRenderProps;
 			});
 			if (error) {
@@ -132,22 +134,28 @@ export default {
 			}
 			if (!renderProps) {
 				return new HtmlWebpackPlugin({
-					filename:        (route + '/index.html').replace(/\/+/, ''),
+					filename:        (route + '/index.html').replace(/^\/+/, ''),
 					template:        'ejs-compiled!./web/index.ejs',
 					googleAnalytics: process.env.GOOGLE_ANALYTICS_ID,
 					title:           '<title>' + process.env.npm_package_name + '</title>',
 				});
 			}
+		<%_ if (state === 'redux') { _%>
 			const store = createStore();
 			const html  = renderToString(<Provider store={store}><RouterContext {...renderProps} /></Provider>);
+		<%_ } else { _%>
+			const html  = renderToString(<RouterContext {...renderProps} />);
+		<%_ } _%>
 			const head  = Helmet.rewind(); // HAS TO be AFTER the renderToString https://github.com/nfl/react-helmet#server-usage
 
 			return new HtmlWebpackPlugin({
-				filename:        (route + '/index.html').replace(/\/+/, ''),
+				filename:        (route + '/index.html').replace(/^\/+/, ''),
 				template:        'ejs-compiled!./web/index.ejs',
 				googleAnalytics: process.env.GOOGLE_ANALYTICS_ID,
 				html:            html,
+		<%_ if (state === 'redux') { _%>
 				state:           JSON.stringify(store.getState()),
+		<%_ } _%>
 				htmlAttributes:  head.htmlAttributes.toString(),
 				title:           head.title.toString(),
 				meta:            head.meta.toString(),
@@ -157,7 +165,7 @@ export default {
 			});
 	<%_ } else if (!render) { _%>
 			return new HtmlWebpackPlugin({
-				filename:        (route + '/index.html').replace(/\/+/, ''),
+				filename:        (route + '/index.html').replace(/^\/+/, ''),
 				template:        'ejs-compiled!./web/index.ejs',
 				googleAnalytics: process.env.GOOGLE_ANALYTICS_ID,
 				title:           '<title>' + process.env.npm_package_name + '</title>',
